pyproject.toml
[tool.poetry]
name = "class-elective-advisor"
version = "0.1.0"
description = ""
authors = ["Matthew Dobley <mkmshopping@sbcglobal.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.4"
langchain-openai = "^0.2.4"
langchain-core = "^0.3.13"
python-dotenv = "^1.0.1"
sphinx = "^8.1.3"


[tool.poetry.group.dev.dependencies]
sphinx-rtd-theme = {version = "3.0.0rc4", allow-prereleases = true}
sphinx-autodoc-typehints = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
